[nomenclature._consensus_maps2consensus_frags] tests refactored here from test_contacts
