+++ tests/test_contacts.py
@@ -2871,35 +2871,3 @@ class Test_modified_mdtraj_contacts(unittest.TestCase):
 
     def test_passes(self):
         contacts._md_compute_contacts.compute_contacts(self.traj, [[0, 1]], scheme='ca', soft_min=True)
-
-class Test_consensus_maps2consensus_frag(unittest.TestCase):
-
-    @classmethod
-    def setUpClass(cls):
-        cls.GPCR = nomenclature.LabelerGPCR(examples.filenames.adrb2_human_xlsx)
-        cls.CGN = nomenclature.LabelerCGN(examples.filenames.CGN_3SN6)
-        cls.geom = md.load(examples.filenames.actor_pdb)
-        cls.maps = [lab.top2labels(cls.geom.top) for lab in [cls.CGN, cls.GPCR]]
-        cls.frags = [lab.top2frags(cls.geom.top) for lab in [cls.CGN, cls.GPCR]]
-
-    def test_works_on_empty(self):
-        maps, frags = _consensus_maps2consensus_frags(self.geom.top, [], verbose=True)
-        assert maps == []
-        assert frags == {}
-
-    def test_works_on_maps(self):
-        maps, frags = _consensus_maps2consensus_frags(self.geom.top, self.maps, verbose=True)
-        self.assertListEqual(maps, self.maps)
-        assert frags == {}
-
-    def test_works_on_Labelers(self):
-        maps, frags = _consensus_maps2consensus_frags(self.geom.top, [self.CGN, self.GPCR], verbose=True)
-        self.assertListEqual(maps, self.maps)
-        for ifrags in self.frags:
-            for key, val in ifrags.items():
-                self.assertListEqual(frags[key],val)
-
-    def test_works_on_mix(self):
-        maps, frags = _consensus_maps2consensus_frags(self.geom.top, [self.maps[0], self.GPCR], verbose=True)
-        self.assertListEqual(maps, self.maps)
-        self.assertDictEqual(frags, self.frags[1])
